Only deploy systemd file when solr v5 or later is used as it is not compatible with v4 or prior.
The template for the systemd script references a bin directory from the
solr_install_path that doesn’t exist in v3 or v4 packages, but is
present in v5 and v6. If the systemd file is put in place it overrides
the init.d script that is also put in place for solr v3 and the systemd
file fails to start the service, where when it does not exist systemd
reverts to using the init.d script via sysv.